refactor owner prompts into its own file
